requirements.txt
asgiref==3.7.2
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1
